api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
Pillow==10.3.0
httpx==0.28.1
pytest==8.2.1
